4.15 - 4.23

+++ part4/blog_list/controllers/blogs.js
@@ -1,16 +1,26 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const {errorHandler} = require('../utils/middleware')
+const {errorHandler, userExtractor} = require('../utils/middleware')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response, next) => {
+blogsRouter.post('/', userExtractor, async (request, response, next) => {
   try {
-    const blog = new Blog(request.body)
+    const user = request.user
+    const blog = new Blog({
+      ...request.body,
+      user: user._id
+    })
+    
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+    
     response.status(201).json(savedBlog)
   } catch (error) {
     next(error)
@@ -30,27 +40,31 @@ blogsRouter.get('/:id', async (request, response, next) => {
   }
 })
 
-blogsRouter.put('/:id', async (request, response, next) => {
-  try {
-    const body = request.body
-    const blog = {
-      likes: body.likes
-    }
-    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
-    response.json(updatedBlog)
-  } catch (error) {
-    next(error)
-  }
+blogsRouter.put('/:id', userExtractor, async (request, response) => {
+  const { title, author, url, likes } = request.body
+
+  const updatedBlog = await Blog.findByIdAndUpdate(
+    request.params.id,
+    { title, author, url, likes },
+    { new: true, runValidators: true, context: 'query' }
+  )
+
+  response.json(updatedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response, next) => {
   try {
-    const result = await Blog.findByIdAndDelete(request.params.id)
-    if (result) {
-      response.status(204).end()
-    } else {
-      response.status(404).end()
+    const blog = await Blog.findById(request.params.id)
+    if (!blog) {
+      return response.status(404).end()
     }
+
+    if (!request.user || blog.user.toString() !== request.user.id.toString()) {
+      return response.status(401).json({ error: 'unauthorized' })
+    }
+
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
   } catch (error) {
     next(error)
   }
